Fixing issues

+++ captum/attr/_core/feature_ablation.py
@@ -433,8 +433,8 @@ class FeatureAblation(PerturbationAttribution):
         current_mask = torch.stack(
             [input_mask == j for j in range(start_feature, end_feature)], dim=0
         ).long()
-        ablated_tensor = (expanded_input * (1 - current_mask).float()) + (
-            baseline * current_mask.float()
+        ablated_tensor = (expanded_input * (1 - current_mask).to(expanded_input.dtype)) + (
+            baseline * current_mask.to(expanded_input.dtype)
         )
         return ablated_tensor, current_mask
